refactor: recompose matrix cache domain
