+++ src/mindroom/matrix/client.py
@@ -24,10 +24,7 @@ from nio.responses import RoomThreadsResponse
 
 from mindroom.constants import STREAM_STATUS_KEY, RuntimePaths, encryption_keys_dir, runtime_matrix_ssl_verify
 from mindroom.logging_config import get_logger
-from mindroom.matrix.cache.event_cache import (
-    ConversationEventCache,
-    normalize_nio_event_for_cache,
-)
+from mindroom.matrix.cache.event_cache_events import normalize_nio_event_for_cache
 from mindroom.matrix.cache.thread_cache_helpers import thread_cache_state_is_usable
 from mindroom.matrix.cache.thread_history_result import (
     THREAD_HISTORY_DEGRADED_DIAGNOSTIC,
@@ -56,6 +53,7 @@ from mindroom.thread_tags import THREAD_TAGS_EVENT_TYPE
 if TYPE_CHECKING:
     from mindroom.config.main import Config
     from mindroom.config.matrix import RoomDirectoryVisibility, RoomJoinRule
+    from mindroom.matrix.cache.event_cache import ConversationEventCache
     from mindroom.matrix.conversation_cache import ConversationCacheProtocol
 
 logger = get_logger(__name__)
